endpoint-to-aboutme

+++ src/controllers/aboutme.controller.ts
@@ -1,12 +1,14 @@
 import { RequestHandler } from "express";
 import { connectionSource } from "../database/data-source";
-import { AboutDetail } from "../database/entity/model";
+import { AboutDetail, User, Section } from "../database/entities";
 import { 
   AboutMeInterface, 
   updateAboutMeInterface 
 } from '../interfaces/aboutme.interface';
 
 const AboutRepository = connectionSource.getRepository(AboutDetail);
+const userRepository = connectionSource.getRepository(User);
+const sectionRepository = connectionSource.getRepository(Section);
 
 
 export const createAboutMe: RequestHandler = async (req, res) => {
@@ -14,38 +16,89 @@ export const createAboutMe: RequestHandler = async (req, res) => {
     // Get the data from the request body
     const { bio, userId, sectionId } = req.body as AboutMeInterface;
 
-    // Check if the interest exists
-    const aboutMeExists = await AboutRepository.findOne({
-      where: { userId },
+    const user = await userRepository.findOne({
+      where: { id: userId },
+    });
+    const section = await sectionRepository.findOne({
+      where: { id: sectionId }
     });
 
-    // If the about me exists, return an error
-    if (aboutMeExists) {
-      return res.status(409).json({
+    if (!user || !section) {
+      return res.status(404).json({
         successful: false,
-        message: "This user already has an about me section created.",
+        message: "User or section does not exist.",
       });
     }
 
-    // Create the interest
+    // Check if the bio exists
+    const aboutMeExists = await AboutRepository.findOne({
+      where: { user },
+    });
+
+    // If the about me exists, delete the bio and create a new one
+    if (aboutMeExists) {
+      await AboutRepository.remove(aboutMeExists);
+
+      const newaboutMe = AboutRepository.create({
+        bio,
+        user,
+        section,
+      });
+      const savedaboutMe = await AboutRepository.save(newaboutMe);
+
+      // Extract the userId, sectionId and sectionName from the savedaboutMe
+      const user_id = savedaboutMe.user?.id;
+      const section_id = savedaboutMe.section?.id;
+      const section_name = savedaboutMe.section?.name;
+      const sectionDetails = { section_id, section_name };
+
+      // Create the data object to be returned
+      const data = {
+        bio: savedaboutMe.bio,
+        user_id,
+        sectionDetails,
+      };
+
+      return res.status(200).json({
+        successful: true,
+        message: "Bio updated successfully",
+        data,
+      });
+    }
+
+    // Create the Bio
     const aboutMeResponse = AboutRepository.create({
       bio,
-      userId,
-      sectionId,
+      user,
+      section,
     });
-    // Save the interest to the database
-    const data = await AboutRepository.save(aboutMeResponse);
+    // Save the About Me to the database
+    const savedaboutMe = await AboutRepository.save(aboutMeResponse);
+    console.log("Saved bio", savedaboutMe);
+
+    // Extract the userId, sectionId and sectionName from the savedaboutMe
+    const user_id = savedaboutMe.user?.id;
+    const section_id = savedaboutMe.section?.id;
+    const section_name = savedaboutMe.section?.name;
+    const sectionDetails = { section_id, section_name };
+
+    // Create the data object to be returned
+    const data = {
+      bio: savedaboutMe.bio,
+      user_id,
+      sectionDetails,
+    };
 
     res.status(201).json({
       successful: true,
-      message: "Interest created successfully",
+      message: "Bio created successfully",
       data,
     });
   } catch (err) {
     console.error(err);
     res.status(500).json({
       successful: false,
-      message: "Could not create interest.",
+      message: "Could not create bio.",
       error: err.message,
     });
   }
@@ -53,17 +106,22 @@ export const createAboutMe: RequestHandler = async (req, res) => {
 
 export const getAboutMe: RequestHandler = async (req, res) => {
   try {
-    const { userId } = req.params;
+    const userId = req.params.userId;
 
-    const aboutMe = await AboutRepository.findOne({ where: { userId } });
+    const user = await userRepository.findOne({ where: { id: userId } });
 
-    if (!aboutMe) {
+    if (!user) {
       return res.status(404).json({ message: 'About me not found' });
     }
 
+    // Retrieve interests from the database for the specific userId
+    const bio = await AboutRepository.findOne({
+      where: { user },
+    });
+
     res.status(200).json({
       successful: true,
-      data: aboutMe,
+      data: bio,
     });
   } catch (error) {
     console.error(error);
@@ -80,25 +138,35 @@ export const updateAboutMe: RequestHandler = async (req, res) => {
     const { userId } = req.params;
     const { bio } = req.body as updateAboutMeInterface;
 
-    const aboutMe = await AboutRepository.findOne({ where: { userId }  });
+    const user = await userRepository.findOne({ where: { id: userId }  });
 
-    if (!aboutMe) {
-      return res.status(404).json({ message: 'About me not found' });
+    if (!user) {
+      return res.status(404).json({ message: 'User not found' });
     }
 
+    // Find the AboutMe entity for the given user
+    const aboutMe = await AboutRepository.findOne({ where: { user } });
+
+    if (!aboutMe) {
+      return res.status(404).json({ message: 'About Me data does not exist' });
+    }
+
+    // Update the AboutMe bio
     aboutMe.bio = bio;
 
+    // Save the updated AboutMe data
     await AboutRepository.save(aboutMe);
 
     res.status(200).json({
       successful: true,
-      message: 'About me updated successfully',
+      message: 'About Me updated successfully',
+      data: aboutMe.bio,  // Return the updated bio
     });
   } catch (error) {
     console.error(error);
     res.status(500).json({
       successful: false,
-      message: 'Could not update about me.',
+      message: 'Could not update About Me.',
       error: error.message,
     });
   }
@@ -106,14 +174,21 @@ export const updateAboutMe: RequestHandler = async (req, res) => {
 
 export const deleteAboutMe: RequestHandler = async (req, res) => {
   try {
-    const { userId } = req.params;
+    const userId = req.params.userId;
+    const user = await userRepository.findOne({ where: { id: userId } });
 
-    const aboutMe = await AboutRepository.findOne({ where: { userId } });
+    if (!user) {
+      return res.status(404).json({ message: 'User not found' });
+    }
+
+    // Find the AboutMe entity associated with the user
+    const aboutMe = await AboutRepository.findOne({ where: { user } });
 
     if (!aboutMe) {
       return res.status(404).json({ message: 'About me not found' });
     }
 
+    // Remove the AboutMe entity from the database
     await AboutRepository.remove(aboutMe);
 
     res.status(200).json({
